requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
croniter==0.3.27
defusedxml==0.5.0
Django==2.1.7
django-anymail==6.0
django-cors-headers==2.5.1
django-environ==0.4.5
django-filter==2.1.0
django-ipware==2.1.0
django-money==0.14.4
django-oauth-toolkit==1.2.0
django-rq==1.3.0
django-starterkit==1.0.4
djangorestframework==3.9.2
Faker==1.0.4
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.1
Pillow==5.4.1
psycopg2-binary==2.7.7
py-moneyed==0.8.0
PyJWT==1.7.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rq==0.13.0
rq-scheduler==0.9
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
stripe==2.21.0
text-unidecode==1.2
urllib3==1.24.1
